ARM: use multi_v7_defconfig by default
versatile_defconfig has been the default defconfig since the beginning of
(git) time, but few people actually have this machine and it uses an
old ARM926 CPU core.

To go with the modern times, this changes the default to multi_v7_defconfig,
which should work on a wide range of modern machines.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>