fs: introduce new 'struct inode_time'
As one part of the puzzle to solve the y2038 problem, this introduces
a new time type to be used for keeping inode timestamps (atime, ctime,
mtime) inside of the kernel.

Initially, this type is defined to 'struct timespec' to allow migrating
all file systems one by one, but the intention is to change the definition
to use either 64-bit signed seconds or 'unsigned long' seconds, which
would allow timestamps between 1970 and 2106.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>

Conflicts:
	fs/inode.c
	include/linux/fs.h
	include/linux/time.h
	kernel/time.c
	kernel/time/timekeeping.c