udf: convert to struct inode_time
udf uses 16-bit year numbers for inode timestamps, which will work
for the 15000 years, but the VFS uses struct timespec for timestamps,
and the Linux udf implementation internally uses a time_t, both of
which are only good until 2038 on 32-bit CPUs.

This gets us one small step closer to lifting the VFS limit by using
struct inode_time in udf, but the implementation still depends on
the 'year_seconds' array that needs to be extended or replaced by
an algorithm that calculates the correct time for every year.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Cc: Jan Kara <jack@suse.cz>