fs: introduce sys_newfstat64/sys_newfstatat64
We want to be able to read file system timestamps beyond year 2038,
which is currently impossibly on 32-bit systems, since none of the
various stat syscall interfaces (oldstat, stat, stat64) handles
correctly.

This introduces a fourth version of the syscalls, called newfstat64
and newfstatat64, which operate on struct newstat64. Each 32-bit
architecture needs to define a version of this structure.
Architectures that have a 64-bit CPU should use the native 64-bit
'struct stat' if possible, so we can avoid adding a compat_newfstatat64
syscall. Note that there is no sys_newlstat64 or sys_newstat64,
as both can be trivially emulated from libc using newfstatat64.

This approach might not be the best solution, as there have been
proposals in the past to add a new 'struct xstat' interface that
would not only solve the y2038 problem but provide a number of
other extensions as well. I have chickened out and avoided reviving
that discussion for now. This new set of syscalls is a much simpler
addition and is hopefully less controversial.
If we end up merging xstat first, we won't need this patch.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>