[SUBMITTED] net: netfilter: fix IPV6 XT sockets dependency
If IP6_NF_IPTABLES=m, and one of NETFILTER_XT_MATCH_SOCKET or
NETFILTER_XT_TARGET_TPROXY is built-in, we get a link error:

net/built-in.o: In function `socket_mt_init':
:(.init.text+0x32a0): undefined reference to `nf_defrag_ipv6_enable'
net/built-in.o: In function `tproxy_tg_init':
xt_TPROXY.c:582: undefined reference to `nf_defrag_ipv6_enable'

The reason for this is that NF_DEFRAG_IPV6 also gets selected
by NF_CONNTRACK_IPV6, and that forces it to be a module.

To solve this, we add another set of dependencies that enforce that
either IP6_NF_IPTABLES is disabled, or the two netfilter drivers
are also loadable modules.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>