gpio: always include linux/gpio/consumer.h from linux/gpio.h
A number of drivers that have started to use the new gpiod interface
are relying on getting the interface declarations from linux/gpio.h or
linux/of_gpio.h, which doesn't work for build-testing on a machine that
isn't using gpiolib at the moment. The problem is that there are
duplicate checks for CONFIG_GPIOLIB: one in consumer.h that provides
an stub API implementation, and one in asm-generic/gpio.h that only
includes the file if gpiolib is enbled.

An example is the tegra DRM driver, which gives us this error
message when built on a platform that doesn't enable GPIOLIB:

drivers/gpu/drm/tegra/output.c: In function 'tegra_connector_detect':
drivers/gpu/drm/tegra/output.c:84:3: error: implicit declaration of function 'gpiod_get_value' [-Werror=implicit-function-declaration]

We could solve this by marking each such driver either 'depends
on GPIOLIB', but that feels like a neverending task because that
configuration tends to not get tested.

Instead, this patch moves the #include vom asm-generic/gpio.h to
the main linux/gpio.h to ensure the interface is always visible,
either as an extern declaration or as a set of stub functions.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>