ARM: move s5pv210 files to mach-exynos
The s5pv210 code is now very similar to the exynos (originally called
s5pc210) code and very small. There are no remaining conflicts between
the two, so it makes sense to consolidate them in one directory.

It is possible to further merge the actual contents as follow-up
patches, for now, this just moves the files without changing them.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>