ARM: exynos: move map*.h headers to mach-exynos
This untangles the map-base.h, map-s5p.h and map-s3c.h header
files and moves the regs-srom.h header so that exynos stops
including most of plat/*.h.

We do get a duplicated S3C_ADDR_BASE definition in the process
that has to match between plat-samsung and mach-exynos, but that
will get fixed in one of the follow-up patches as we further
separate the two directories.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>