[REWRITE] vfio/pci: make MSI handling optional
A recent bug fix to the MSIx handling in vfio added references
to functions that may not be defined if MSI is disabled in the kernel,
resulting in this link error:

drivers/built-in.o: In function `vfio_msi_set_vector_signal':
:(.text+0x450808): undefined reference to `get_cached_msi_msg'
:(.text+0x45080c): undefined reference to `write_msi_msg'

From what I can tell, all the MSI specific ioctl handling can
be made optional in this case, which also reduces the code size
for the vfo driver. This patches does so using a build-time
IS_ENABLED() check, which leaves all the build coverage testing
present but avoids the link error.

A side-effect of this is that the driver now returns -ENOTTY instead
of -EINVAL if user space calls VFIO_PCI_MSI_IRQ_INDEX on a kernel
without MSI support, which seems to be the best behavior, but
the approach could easily be changed if we want to preserve the
existing behavior for compatibility reasons.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: b8f02af096b1 ("vfio/pci: Restore MSIx message prior to enabling")
Link: https://lkml.org/lkml/2014/10/13/124