ARM: at91: simplify sama5 machine init code
The sama5 initialization can be simplified a lot as we already know
from the compatible string what SoC type we are running on and we
can just initialize that and print the subtype as detected.

The at91_init_soc handling can be completely eliminated this way,
reducing both source and object code size.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>