ARM: at91: remove last bits of at91_init_soc
This removes the init_soc abstraction level in the sam9 code, by
open-coding the differences in the map_io and init_early callbacks.
As the differences are very minor, we can check for the specific
sam9 soc directly in the map_io and init_early functions, without
going through an indirect function call.

Further simplification is possible if the rtt/rtc drivers learn
to mask the irq themselves and if we detect the SRAM location
from DT.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>