ARM: orion5x: make PCI support a platform driver
This adds the necessary glue code to make the PCI (non-PCIe) section
arch/arm/mach-orion5x/pci.c a standalone platform driver. The existing
probing method is kept alive, but board files can choose to use the
new driver now.

PCI is hardcoded to use domain 1, and the second I/O space window,
to enable coexisting with PCIe. Boards that register the PCI host
controller as a platform device have to ensure that hw_pci.nr_controllers
is '1' now, so it does not get registered twice.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>