ARM: orion5x: remove dependency on hardcoded constants
As a preparation for DT support, this changes the orion5x PCI driver
further to no longer depend on global variables or constants outside
of the probe function. Instead we now use a data structure containing
the private data of this driver. This is allocated and filled by
the probe function, and will be filled with data from DT in the next
step.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>