ARM: orion5x: DT: enable PCI nodes
The PCIe node does not require any board specific configuration
and just needs to be enabled. The PCI node needs a custom
interrupt-map because the IntA and IntB pins of the only
slot are connected to GPIOs.

The IDSEL pin 7 that is used on this board corresponds to
device address 0x3800 0 0, and we map IntA (1) to gpio 7
and IntB (2) to gpio 6, as in the board file.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>