[SUBMITTED] SCSI: ips: remove pointless #warning
The ips driver contains

 #warning "This driver has only been tested on the x86/ia64/x86_64 platforms"

which gets printed by the compiler every time this driver gets
build for something other than those three architectures. The same
is true for most other drivers, but none of them prints a warning like
this.

Let's remove the warning now to reduce the amount of clutter.

FWIW, I've reviewed the driver and found a few oddities, but nothing
that needs to be fixed or that would be architecture specific.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>