[DUPLICATE] ARM: sunxi: always select RESET_CONTROLLER
A lot of drivers for mach-sunxi depend on RESET_CONTROLLER. While
this is selected for the drivers that have it, we keep running
into build errors in drivers that are used on both the SoCs with
and without a reset controller.

This patch moves the select statements into the common sunxi
Kconfig option, which solves multiple such problems.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Link: http://lists.infradead.org/pipermail/linux-arm-kernel/2015-February/322806.html
Cc: Maxime Ripard <maxime.ripard@free-electrons.com>