ARM: orion: move watchdog setup to mach-orion5x
The watchdog device node is created in plat-orion/common.c
but depends on the bridge address that is platform specific,
so as a preparation for orion multiplatform support, we should
move it out of the common code into orion5x.

The other two users of plat-orion currently do not support
the watchdog device, so this is not a problem.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Acked-by: Andrew Lunn <andrew@lunn.ch>