mmc: msm: move data mover into mmc driver
The msm_sdcc driver is the only user of the proprietary msm
data mover interface. In order to simplify the multiplatform
conversion, this patch moves the entire data mover code into
the mmc driver.

If against our expectations any other drivers should be added
in the future that use the same data mover, the best approach
would be to convert this code into a dmaengine driver and use
the generic interfaces for that.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>