mmc: msm: pass dmov resources via device
The data mover code that got moved here from arch/arm was not
using the device model at all. This tries to retrofit it
passing the irq and mmio address using resources.

The approach is a little half-baked, but it still seems better
than the original code. One problem now is that the resources
are not getting cleaned up properly. Instead of doing something
that I can't test myself, I just disable module unloading to
avoid having to rewrite more than necessary.

It's unclear whether this works with multiple sdcc devices.
What I know is that no platform in the kernel registers more
than one at the moment, and I suspect that trying to use more
than one was already broken. If it was working previously,
it most likely keeps working, and it's unlikely that we will
see new users of this driver that require more than one
device.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>