mmc: msm: move clk-reset logic to platform
There is no generic interface for resetting a clock, and
creating a reset driver for msm seems overkill, so this
moves the reset logic from the msm_sdcc driver into
a platform_data callback that calls into the clock driver.

This follows the model that is used for all other devices
on the msm platform that require a clk reset.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>