ARM: msm: fix mach/msm_iomap.h inclusions
The mach/msm_iomap.h header is broken for files that need to
be built for multiple SoCs, because it includes mutually
exclusive headers for each variant.

This removes the #ifdef and the multiplexer and includes the
correct files in all places directly. To do this, we also have
to split up the io.c file and move the contents into the per-soc
files.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>