ARM: msm: make smd behave like a normal driver
The smd driver is registered as a platform driver with a probe
function, but hardcodes soc-specific interrupts and other settings
at compile-time, which gets in the way of reusing the same driver
on multiple machines.

This changes all compile-time configuration in the smd driver
to runtime configuration from platform device resources and
the platform device id, like we do for other drivers.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>