[SUBMITTED] [media] vb2: remove unused variable
A recent bug fix removed all uses of the 'fileio' variable in
vb2_thread_stop(), which now causes warnings in a lot of
ARM defconfig builds:

drivers/media/v4l2-core/videobuf2-core.c:3228:26: warning: unused variable 'fileio' [-Wunused-variable]

This removes the variable as well. The commit that introduced
the warning was marked for 3.18+ backports, so this should
probably be backported too.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Fixes: 0e661006370b7 ("[media] vb2: fix 'UNBALANCED' warnings when calling vb2_thread_stop()")
Cc: <stable@vger.kernel.org>      # for v3.18 and up