[SUBMITTED] HSI: cmt_speech: fix timestamp interface
The user interface for timestamps in the new cms_speech
driver is broken in multiple ways:

- The layout is incompatible between 32-bit and 64-bit user
  space, because of the size differences in 'struct timespec'.
  This means that the driver can not work when used with 32-bit
  user space on a 64-bit kernel.

- As there are plans to change 32-bit user space to use
  a 64-bit time_t type in the future, it will also be
  incompatible with new 32-bit user space.

As the driver is not yet widely used, now would be a good time
to change it to just use a 64-bit nanosecond value, which also
happens to be more efficient. In order to make the layout
non-ambiguous, we have to also add extra padding so the 64-bit
value is naturally aligned.

The change requires the respective update to user space tools
using it. If that is for some reason not possible any more,
another solution would be to replace the timespec structure
with two __u32 values to avoid incompatibilities later.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>