y2038: introduce CONFIG_COMPAT_TIME
There are a total of system calls (aside from ioctl) that pass a time_t
or derived data structure as an argument, and in order to extend time_t
to 64-bit, we have to replace them with new system calls and keep providing
backwards compatibility.

To avoid adding completely new and untested code for this purpose, we
introduce a new CONFIG_COMPAT_TIME symbol that is always set on 64-bit
architectures with 32-bit compat mode, as well as all 32-bit architectures
that have been extended to provide the new system calls.

After this is done for all architectures, the CONFIG_COMPAT_TIME symbol
can be made a user-selected option, to enable users to build a kernel
that only provides y2038-safe system calls.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>