y2038: introduce struct __kernel_timespec
A lot of system calls pass a 'struct timespec' from or to user space,
and we want to change that type to be based on a 64-bit time_t by
default.

This introduces a new type struct __kernel_timespec, which has the
format we want to use eventually, but also has an override so all
architectures that do not define CONFIG_COMPAT_TIME yet still get the
old behavior.
Once all architectures set this, we can remove that override.

This also introduces a get_timespec64/put_timespec64 set of functions
that convert between a __kernel_timespec in user space and a timespec64
in kernel space.

The current behavior of get_timespec64 explicitly zeroes the upper half
of the tv_nsec member, to allow user space to define its own 'struct
timespec' with some padding in it. Whether this is a good or bad idea
is open for discussion.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>