y2038: add compat_sys_rt_sigtimedwait variants
rt_sigtimedwait takes sigevent and sigset arguments that are always
different between 32-bit and 64-bit user space. It also takes a
timespec argument that will be the same in the future between
architectures, and we want to share the compat system call handling.

We need to support all four combinations, and this adds the two new
ones:

- native sys_rt_sigtimedwait: now takes a __kernel_timespec that
  will at some point use 64-bit time_t on all architectures
- compat_sys_rt_sigtimedwait on 64-bit kernels: emulates the
  behavior of the old 32-bit user space
- compat_sys_rt_sigtimedwait on 32-bit kernels: emulates the
  old timespec but uses the native 32-bit sigevent and siginfo
  arguments
- compat_sys_rt_sigtimedwait_time64 on 64-bit kernels: emulates
  the 32-bit sigevent and siginfo structures but uses the native
  __kernel_timespec.

Having separate code for all four variants is a bit ugly, but
I could not come up with a better solution, and the code duplication
is fairly low in the end.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>