y2038: introduce compat_sys_pselect6_time64
As 32-bit architectures now have two variants of pselect6, one using
32-bit time_t and one using 64-bit time_t, we unfortunately also need
to add a second compat case for the version of pselect6 that now
uses a 64-bit time_t but the 32-bit format for fd_set and sigset,
which is different from what a 64-bit kernel uses natively.

This is complicated by the way that poll_select_copy_remaining()
works, which already copies back the remaining time as either
timespec or timeval, depending on whether it is used by
select or pselect6. We extend it in a similar fashion to also
copy back a __kernel_timespec.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>