y2038: introduce struct __kernel_timex
This is one of multiple possible extended definitions for timex, to
allow future user space to keep working while defining time_t
to 64-bit. The idea is that clock_adjtime() will take the new
structure, which has an identical layout between 32-bit and 64-bit
tasks, so we can use the compat_sys_clock_adjtime() function to
always handle the old 32-bit layout, and use sys_clock_adjtime()
to handle the new layout, without introducing an extra compat
function.

Alternatives to this approach are:

- define a new __kernel_timex that only differs in the size of the
  timex->time member, so we can keep using the old definition in
  source form, but with the new timeval definition. This requires
  adding an extra system call on 64-bit machines, to handle all
  three layouts.

- use some of the padding words to store the upper half of the
  time_t. This allows us to keep using the existing system calls,
  but requires source-level changes in every user program to
  actually access the new data.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>